Moved `method` routes to their own blueprint
