+++ mycodo/mycodo_flask/methods/__init__.py
@@ -1,0 +1,3 @@
+# coding=utf-8
+""" method endpoints """
+from . import views
